shared/requirements.txt
praw
PyYAML